refactor for clarity and simplicity (broke alternating turns tho)
