+++ tools/ui_widgets.py
@@ -1,21 +1,18 @@
 '''Used for easily initializing widgets, meant to reduce the repeated calls which specify the respective fields of similar widgets'''
 # pylint: disable=C0321, no-name-in-module, too-many-arguments
-from enum import Enum
 from typing import List, Tuple, Optional, Union
 from PyQt5.QtWidgets import QPushButton, QLabel, QGroupBox, QGridLayout, QVBoxLayout, QToolButton, QLineEdit, QTextEdit, QRadioButton
 from PyQt5.QtGui import QValidator
 from PyQt5.QtCore import Qt
+from tools.stylesheets import MACRO_BTN_STYLE
 
-class SquareState(Enum):
-    '''Define control of squares'''
-    PLAYER_A = 1
-    PLAYER_B = -1
-    PLAYER_NONE = 0
-
-class Turn(Enum):
-    '''Define whose turn it is'''
-    IS_PLAYER_ONE = 1
-    IS_PLAYER_TWO = 0
+class MenuButton(QPushButton):
+    """A standardized button for the game"""
+    def __init__(self, text: str = '', layout_position: Tuple[int,...] = (0,0)):
+        super().__init__(text = text)
+        self.setStyleSheet(MACRO_BTN_STYLE)
+        self.setFixedSize(160, 40)
+        self.layout_position = layout_position
 
 class GameButton(QPushButton):
     """A standardized button for the game"""
@@ -26,21 +23,14 @@ class GameButton(QPushButton):
         if height: self.setFixedHeight(height)
         self.layout_position = layout_position
 
-class WAAMToolButton(QToolButton):
-    """A standardized button for the WAAM GUI"""
-    def __init__(self, increment_key: str, continuous_key: str, text: str = '', style: str = '', width: Optional[int] = None, height: Optional[int] = None,
-                       layout_position: Tuple[int,...] = (0,0), arrow: Optional[Qt.ArrowType] = None, multiplier: int = 1, is_rotation: bool = False):
+class GameGroupBox(QGroupBox):
+    """A standard groub box widget for the WAAM GUI"""
+    def __init__(self, style:str, width: Optional[int] = None, height: Optional[int] = None):
         super().__init__()
-        self.setText(text)
-        self.increment_key = increment_key
-        self.continuous_key = continuous_key
         self.setStyleSheet(style)
+        self.setLayout(QGridLayout())
         if width: self.setFixedWidth(width)
         if height: self.setFixedHeight(height)
-        if arrow: self. setArrowType(arrow)
-        self.layout_position = layout_position
-        self.multiplier = multiplier
-        self.is_rotation = is_rotation
 
 class GameLabel(QLabel):
     """A standardized label for the game"""
@@ -69,15 +59,6 @@ class WAAMTextEdit(QTextEdit):
         if width: self.setFixedWidth(width)
         if height: self.setFixedHeight(height)
         self.layout_position = layout_position
-
-class WAAMGroupBox(QGroupBox):
-    """A standard groub box widget for the WAAM GUI"""
-    def __init__(self, style:str, layout: Union[QGridLayout, QVBoxLayout], width: Optional[int] = None, height: Optional[int] = None):
-        super().__init__()
-        self.setStyleSheet(style)
-        self.setLayout(layout)
-        if width: self.setFixedWidth(width)
-        if height: self.setFixedHeight(height)
 
 class WAAMRadioButton(QRadioButton):
     """A standard radio button for the WAAM GUI"""
